implement rest api sebelum menggunakan midtrans

+++ migrations/000002_produk.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE produks (
     id SERIAL PRIMARY KEY,
     name VARCHAR(255) NOT NULL,
-    price DECIMAL(10, 2) NOT NULL,
+    price int NOT NULL,
     stock_quantity INT NOT NULL
 );
