+++ migrations/000004_order_item.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE order_items (
     pesanan_id INT,
     produk_id INT,
     quantity INT NOT NULL,
-    total DECIMAL(10, 2) NOT NULL,
+    total int NOT NULL,
     FOREIGN KEY (pesanan_id) REFERENCES pesanans(id),
     FOREIGN KEY (produk_id) REFERENCES produks(id)
 );
